requirements.txt
aiogram>=3.7
aiohttp
pydantic>=2.0
pandas
python-dotenv
pydantic-settings